witx: change identifiers to snake case
Change all $SHOUTY_SNAKE_CASE identifers to $snake_case.

It is the responsibility of a code generator backend to select the
right case and style of these identifiers for the target language.
In witx, we should standardize on snake case (lowercase with a
single underscore as a word separator)

sed -i -E 's/\$([A-Z0-9_]+)([)\n\r ])?/\$\L\1\2/g' **/*.witx